Refactor current methods
