+++ app/controllers/concerns/spina/current_theme.rb
@@ -1,0 +1,17 @@
+module Spina
+  module CurrentTheme
+    extend ActiveSupport::Concern
+    
+    included do
+      before_action -> { current_theme }
+      helper_method :current_theme
+    end
+  
+    private
+    
+      def current_theme
+        Spina::Current.theme ||= ::Spina::Theme.find_by_name(current_account.theme)
+      end
+        
+  end
+end
