+++ app/controllers/spina/application_controller.rb
@@ -1,6 +1,6 @@
 class Spina::ApplicationController < Spina.frontend_parent_controller.constantize
   include Spina.config.authentication.constantize
-  include Spina::CurrentMethods
+  include Spina::CurrentAccount, Spina::CurrentTheme
   
   helper Spina::Engine.helpers
